refactor: 코드 간결화 및 컴포넌트 분리
