+++ src/app/(with-nav-footer)/(mypage)/my-activities/[id]/reservation-dashboard/_components/MyActivitiesReservations.tsx
@@ -4,11 +4,12 @@ import CloseImage from '@/assets/icons/close.svg';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { useReservedSchedule, useReservations } from '@/lib/hooks/useMyActivities';
 import { useClickOutside } from '@/lib/utils/useClickOutside';
-import FilterDropdown from '@/components/FilterDropdown';
 import ReservationCardList from './ReservationCardList';
-import arrowFilterDropdown2 from '@/assets/icons/arrow-filter-dropdown2.svg';
 import Image from 'next/image';
-import NoData from '@/assets/icons/No-data.svg';
+import ReservationStatusTabs from './ReservationStatusTabs';
+import NoReservations from './NoReservations';
+import ReservationsTimeSelect from './ReservationsTimeSelect';
+import LoadingSpinner from '@/components/LoadingSpinner';
 
 type Props = {
   selectedDate: string;
@@ -22,7 +23,7 @@ type FilterDropdownOption = {
   onClick?: () => void;
 };
 
-type ReservationStatus = 'pending' | 'confirmed' | 'declined';
+export type ReservationStatus = 'pending' | 'confirmed' | 'declined';
 
 export default function MyActivitiesReservations({ selectedDate, setSelectedDate, activityId }: Props) {
   const {
@@ -40,14 +41,6 @@ export default function MyActivitiesReservations({ selectedDate, setSelectedDate
   const [isOpen, setIsOpen] = useState(false);
   const popupRef = useRef<HTMLDivElement | null>(null);
   const reservations = reservationsData?.reservations || [];
-
-  useEffect(() => {
-    console.log('Fetching reservations with:', {
-      activityId,
-      scheduleId: selectedSchedule?.value,
-      status: activeTab,
-    });
-  }, [activityId, selectedSchedule, activeTab]);
 
   const options = useMemo(() => {
     return dateSchedule
@@ -79,11 +72,6 @@ export default function MyActivitiesReservations({ selectedDate, setSelectedDate
     setSelectedDate('');
   });
 
-  const handleClickClose = () => {
-    setSelectedDate('');
-    setIsOpen(false);
-  };
-
   const handleSelect = (option: FilterDropdownOption | null) => {
     setSelectedSchedule(option);
   };
@@ -104,13 +92,9 @@ export default function MyActivitiesReservations({ selectedDate, setSelectedDate
     }
   }, [filteredSchedule, selectedSchedule, activeTab, refetchReservations]);
 
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
+  if (isLoading) return <LoadingSpinner />;
 
-  if (isError) {
-    return <div>Error loading reservation data.</div>;
-  }
+  if (isError) return <div>에러가 발생했습니다</div>;
 
   return (
     <div ref={popupRef}>
@@ -119,7 +103,10 @@ export default function MyActivitiesReservations({ selectedDate, setSelectedDate
           <div className='flex items-center justify-between'>
             <h1 className='text-[24px] leading-[32px] font-bold'>예약 정보</h1>
             <Image
-              onClick={handleClickClose}
+              onClick={() => {
+                setSelectedDate('');
+                setIsOpen(false);
+              }}
               className='cursor-pointer'
               src={CloseImage}
               width={30}
@@ -128,43 +115,20 @@ export default function MyActivitiesReservations({ selectedDate, setSelectedDate
               aria-label='예약 현황창 닫기'
             />
           </div>
-          <div>
-            <div className='mb-4 text-[20px] font-semibold'>예약 날짜</div>
-            <div className='mb-1 text-[20px] leading-[32px] font-normal'>{formattedDate}</div>
-            <FilterDropdown
-              options={options}
-              onSelect={handleSelect}
-              label='시간대 선택'
-              icon={arrowFilterDropdown2}
-              selected={selectedSchedule || { label: '시작시간 ~ 종료시간', value: '' }}
-              buttonClassName='w-full rounded-[4px] h-[56px] border border-gray-800 py-2 px-4 text-black-100 bg-white'
-              optionClassName='border-gray-800 p-3'
-              dropdownClassName='w-full border rounded-[4px] border-gray-800 overflow-y-auto overflow-x-auto bg-white'
-            />
-          </div>
+          <ReservationsTimeSelect
+            formattedDate={formattedDate}
+            options={options}
+            selectedSchedule={selectedSchedule}
+            handleSelect={handleSelect}
+          />
           {filteredSchedule && filteredSchedule.length > 0 ? (
             <>
               <div className='h-[500px] rounded-b-[24px] bg-gray-100'>
-                <section className='flex items-center gap-5 border-b border-b-gray-300 bg-white text-[20px] leading-[32px] font-normal text-gray-900'>
-                  <button
-                    className={`cursor-pointer ${activeTab === 'pending' ? 'pm-[6px] border-b-[4px] border-green-100 text-[20px] leading-[32px] font-semibold text-green-100' : ''}`}
-                    onClick={() => setActiveTab('pending')}
-                  >
-                    신청 {filteredSchedule[0].count.pending}
-                  </button>
-                  <button
-                    className={`cursor-pointer ${activeTab === 'confirmed' ? 'pm-[6px] border-b-[4px] border-green-100 text-[20px] leading-[32px] font-semibold text-green-100' : ''}`}
-                    onClick={() => setActiveTab('confirmed')}
-                  >
-                    승인 {filteredSchedule[0].count.confirmed}
-                  </button>
-                  <button
-                    className={`cursor-pointer ${activeTab === 'declined' ? 'pm-[6px] border-b-[4px] border-green-100 text-[20px] leading-[32px] font-semibold text-green-100' : ''}`}
-                    onClick={() => setActiveTab('declined')}
-                  >
-                    거절 {filteredSchedule[0].count.declined}
-                  </button>
-                </section>
+                <ReservationStatusTabs
+                  activeTab={activeTab}
+                  setActiveTab={setActiveTab}
+                  count={filteredSchedule[0].count}
+                />
 
                 {activeTab && selectedSchedule && (
                   <div className='px-2 pt-[27px]'>
@@ -174,15 +138,7 @@ export default function MyActivitiesReservations({ selectedDate, setSelectedDate
               </div>
             </>
           ) : (
-            <div className='rounded-b-[24px] bg-gray-100'>
-              <div className='h-[1px] w-full bg-gray-300'></div>
-              <div className='flex flex-col items-center gap-4 px-[50px] py-[145px]'>
-                <Image src={NoData} width={80} height={80} alt='선택된 시간대에 대한 예약 정보가 없습니다.' />
-                <div className='items-center justify-center text-center text-gray-500'>
-                  선택된 시간대에 대한 예약 정보가 없습니다.
-                </div>
-              </div>
-            </div>
+            <NoReservations />
           )}
         </div>
       )}
